Asset system overhaul
Simplified things vastly for now. Got rid of the asset store stuff, now
we have a global Filesystem class (instance stored in Engine) that is
used by the asset system (and anything else that needs FS). Initially it
just uses the real FS, but in future there will be a layered implementation
that can layer package files on top of the real FS transparently to users.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>