Replace DynamicUniformBuffer with new UniformBuffer classes
UniformBufferBase is a non-template base class which has an associated
UniformStruct descriptor, and has methods to generically set buffer
members by name. It also maintains a CPU shadow buffer and (crudely,
for now) tracks modifications to the buffer to determine when to
upload changes to the GPU.

UniformBuffer<Struct> is a template version for a specific struct type,
and allows direct access to the buffer content.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>