GL: Clear cached texture binding state on texture deletion
If a texture is deleted that is currently bound to some unit, it will remain
set as the bound texture in the state cache. Should a new texture then be
created which is assigned the same object ID and then bound, it will not be
bound if the cache still refers to the old texture.

Fix by invalidating the state cache entry on deletion of a texture.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>