Rename GPUShader to GPUProgram
This is to avoid confusion between terminology. Theare are Shader assets
which refer to a whole combination of passes and programs for each stage
within those. Shader was also used to refer to an individual program for
a single stage on the GPU layer. Renaming makes the two distinct concepts.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>