Handle Entity scale in CollisionShape
Bullet provides a setLocalScaling() method for shapes that support non-
uniform scaling, but not for ones that don't (sphere, capsule). Therefore
to implement scale we just re-create the shape when the scale changes,
and bake the scale into the shape dimensions. Sphere and capsule will
raise an error if they are given a non-uniform scale.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>