Initial work on object metadata system
Not usable yet, just some work I've been doing outside of the repository
that I want to get into version control.

This is mainly the initial work on objgen, a tool that will be used to
generate metadata based on annotations in source code. This will be used
for type introspection at runtime which will allow for features such as
dynamic class properties and method invocation.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>