Finished ImGUI integration for debug overlay
Added input handling, and an API for adding windows into the debug GUI.
The debug GUI has a top-level menu which allows all windows registered
to be opened. The debug GUI is disabled by default, it can be enabled by
pressing F1, which will make it visible and capture input. Pressing F2
will toggle input capturing, so control can be returned to the game
while still keeping the GUI visible.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>