Allow custom textures to be used with ImGUI
ImGUI has image methods which allow a custom texture to be rendered. To
do this a void * handle is passed into the method which comes out in the
draw list. To allow this to be used with low-level GPU textures, the
rendering code has been changed to use GPU textures instead of a texture
asset. Passing a texture to ImGUI requires a GPUTexture or Texture2D to
be passed through DebugWindow::refTexture, which adds it to a list of
GPUTexturePtrs which is held until after the overlay has been rendered,
ensuring that the texture will not be destroyed before it is used.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>