Remove EngineGlobal, create DebugManager early
It was supposed to try to implement a deterministic destruction order
for global objects by destroying them in the order in which they are
created, however there is now a case where that is not right. We want
the debug manager to be created early, so that other systems can
register debug windows with as part of their initialisation, but the
debug manager must be destroyed in a different to free some of its
resources before their managers are destroyed.

Therefore, just remove EngineGlobal and instead handle destruction of
global systems manually in Engine's destructor.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>