Render non-font textures correctly on debug overlay
The font atlas was stored in a single-channel R8 texture, and the debug
overlay shader sampled that channel and moved it to the alpha channel,
setting all other channels to 1. This is not correct for other textures
as we want to preserve all of their channels. As a (non-ideal) fix, make
the font atlas an RGBA texture instead. The better solution to do in
future would be to make use of texture swizzling parameters to handle
the swizzling needed on a single-channel font atlas outside of the shader.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>