Add an "asset explorer" debug window
This allows to inspect the currently loaded assets, and view some details
about them. Asset types can implement an explore() method to add elements
to the GUI. This is currently implemented for TextureBase, to display
some texture details and display the texture contents. This will probably
change a bit after the object system is implemented.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>