A few object cleanups
Refactored annotation handling in objgen slightly, we now will error if
a class annotation is placed on the wrong variable. Also made the CLASS()
macro not change visibility at all, it should now be placed in the public
part of the class, and this means we can also remove the STRUCT() macro.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>