Add a system for basic metadata about non-Object-derived types
The object property system and dynamic method invocation will require
some way to get some basic information about types that are not Object-
derived classes (and therefore have not had code generated for them by
objgen), in order to do type checking.

This commit implements a system to do this, by using some fun magic to
dynamically generate such metadata at runtime as it is required. A new
class, MetaType, is added which represents a basic type. MetaClass now
derives from this. Type comparisons can be done by comparing MetaType
pointers. MetaTypes are created as needed through a template method,
MetaClass::lookup(). This uses a static local to cause the MetaType to
be allocated the first time it is called for a given type. Type names
are determined by a wonderful hack: extract it from the string given
by __PRETTY_FUNCTION__. There is a specialisation for Object-derived
classes which uses the class' static MetaClass instead.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>