Implement "virtual properties"
It may sometimes be desirable to have class properties which do not
directly correspond to a member variable. This allows, for instance, a
property to be exposed as one type for the purpose of editing it, but
the internal representation of it is a different type.

This is initially used for the transformation properties of Entity:
internally, these are stored inside a Transform instance, but we want
the individual parts of the transformation to be exposed as properties
on Entity so they can be edited separately.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>