Some work on a "world explorer"
This allows the state of the world to be examined and properties on
entities and components to be viewed and modified.

This is almost getting towards a world editor, however it's sort of a
temporary solution at the moment. The real editor would be implemented
outside of the engine itself.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>