Add default getter/setter names for virtual properties, remove private properties
If no getter/setter methods are specified for virtual properties, a default
name is now chosen based on the property name, e.g. for a property named
"position", the getter is "position" and the setter is "setPosition".

Also, remove support for private properties. In cases where these would be
used virtual properties should be used instead.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>