Ensure that errors in virtual property declarations are reported properly
Because VPROPERTY() expanded to nothing outside of objgen, and because
during the build objgen does not report libclang diagnostics so that any
errors are reported through the real compiler, errors in virtual property
declarations such as bad types were not reported. The only error observed
would be link errors due to the missing reflection code.

Fix this by always declaring the static member variable for the property
even during normal compilation. The variable is never actually defined so
this has no real effect, and allows the real compiler to report errors in
virtual property declarations.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>