Implement construction of objects through the object system
This allows an object to be constructed given a MetaClass for the
intended type of the object. This means that it is possible, for example,
to construct an object given a string containing the name of a class.
This will be used both for the creation of components in the editor, and
as part of the serialization system.

Only constructors with no parameters are supported. A class which does
not have a suitable constructor is marked as non-constructable. A class
which has a suitable but non-public constructor cannot be constructed
through the public interface, however that constructor will still be
usable for serialization.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>