objgen: Improve handling of clang errors during build
When run as part of the build, a flag to ignore clang errors is passed
to objgen. When a clang error occurs with this flag, objgen silently
generates an empty output file. The intention of this is to let errors
in the code be reported by the real compiler, as that can give much
better diagnostics than are available through libclang.

However, a few cases were just encountered where an error occurs when
parsing a source file with objgen but not in the real build, due to
missing includes in the header file being processed. Such failures only
appear as link errors later on due to the missing objgen output.

To fix this, instead of generating a completely empty output file when
a clang error is encountered, generate one which only includes the
offending source file. This should ensure that any errors that occur
when objgen parses the file are reported when the real compiler tries
to compile the objgen output.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>