Implement a better interface between the engine and game code
Previously a game namespace was defined that was expected to contain
functions to get the engine configuration and to create an instance of
the Game class.

Implemented a cleaner solution using the object system: everything is
now inside the Game class (which now derives from Object) and game code
is expected to define a class derived from this. Early on in engine
initialisation the object system is used to find and construct an
instance of this class.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>