Add information about enum constants to MetaType
If the type of a class property is an enum, objgen will now generate
metadata about the enum type including the name and value of all of its
constants. This gets associated with the dynamically-generated MetaType
instance for the enum type, which allows the names of enum property
values to be retrieved at runtime.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>