objgen: Fix code generation for namespaced/nested classes/enums
We name some of the variables in the generated code based on the class
or enum name. If the class or enum resides in a namespace or nested in
a class, this generated an invalid variable name, e.g. "Foo::Bar_data".
Fix by replacing "::" with "_" in the names for such variables.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>