WorldExplorer: Add a custom editor for MeshRenderer's materials list
It's not really possible to handle the materials list through the property
system. For cases like these it's fine to have some custom editing logic,
so implement some so that we can modify the materials list through the
world explorer.

With this, it's now possible to completely build a scene at from the
world explorer at runtime!

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>