Add serialisation support to the world system
A World instance can now be serialised, which will serialise all of the
entities it contains and their components. Deserialising is possible and
works correctly, however there is currently no interface for doing so (I
hacked in some code to test). Some work needs to be done on the asset
system for this to be done properly.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>