Add support for loading serialised assets
This adds a new asset loader which treats any asset with the extension
".object" as a JSON-serialised asset object. In the short-term the goal
is to replace AssetLoader::dataIsMetadata() with the use of serialised
objects, since this is used for "native" asset types. Longer term, at
game runtime the only thing that will be supported is loading serialised
objects. The task of importing other file formats will be moved out to
the editor or some other tool.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>