Vulkan: Remove unnecessary state object creation stubs
The blending, depth/stencil, rasterization and vertex input states all
form part of a pipeline object in Vulkan. Everything to do with them
will be handled by the pipeline management code when that is written,
we do not need any specialisations of these over the generic versions.

Since GPUManager ensures that only one state object is ever created
for any given state description, later on we will use pointers to the
state objects to form part of a lookup key in a cache of pipeline
objects. If a new pipeline is required for a given state, then the
necessary Vulkan state information structures will be filled in from
the generic state descriptors at that point.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>