Define new GPU resource set API
This will replace the current GL-like binding model with something that
maps better to Vulkan's descriptor set model. By exposing Vulkan-like
functionality for this in the generic API we can create and maintain
descriptor sets up front rather than having to build them as needed in
the Vulkan backend. This model can be emulated on top of GL easily.

This commit adds the initial definition of the API and begins to
implement it for the GL backend, but there is no usage of it yet.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>