New shader compilation pipeline based on glslang/SPIR-V
Instead of handing GLSL code directly to the GPU backend, we now have
our own compiler based on glslang, the Khronos reference GLSL compiler.
This compiles to SPIR-V and passes that to the GPU backend. For the
OpenGL backend, we're compiling this back to GLSL for the driver using
SPIRV-Cross.

There are multiple reasons why this has been done. Firstly, Vulkan
requires the use of SPIR-V shaders, so we need our own GLSL compiler
for that. It makes sense to take it and use it to give us a unified
target to write our shaders for, rather than just keeping it exclusive
to the Vulkan backend. Furthermore, since the engine's view of the
world is becoming more Vulkan-like, we want to be able to write our
shaders accordingly, e.g. using features like descriptor sets. This
allows us to do this, and we can transform Vulkan-like features to work
with GL using SPIRV-Cross.

Note that for now we are still doing the whole shader compilation from
the original GLSL source at runtime. The intention is that this will
change in future. A "compiled" game would probably include only the
SPIR-V binaries for its shader variations in its shader assets with no
source, and the game binary would not include the shader compiler at
all.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>