Register assets in AssetManager prior to calling deserialise() on them
This is done for 2 reasons. Firstly it allows deserialise() to know the
asset path, which is used by the next commit. Secondly, it means that
if a serialised asset or any child objects it contains have a reference
back to it, it will be correctly resolved to the asset rather than
recursively attempting to load it.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>