Implement render passes in the GPU layer
This adds Vulkan-style render passes to the generic GPU interface. This
will allow us to directly use this functionality in the Vulkan backend
without additional guesswork to figure out where render passes start
and end. This is easily mapped onto GL.

For now we don't expose subpasses through the generic interface, as
using subpass inputs in Vulkan requires special handling in shaders
which would need translating to work on GL. This can be added later.
Also, the use of render passes in the main renderer is a bit messy at
the moment, but I intend to refactor the renderer after I've done the
main changes to the GPU layer for Vulkan.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>