Add support for GL debug groups
This adds support for GL_KHR_debug's debug group facility, and adds debug
groups throughout the rendering code. Debug groups are added right down
to individual SceneEntities and SceneLights. This makes it far, far easier
to search through a frame in apitrace, and I don't really know why I've
not added this before.

As part of this a name was added to SceneEntity and SceneLight, which is
kind of tacked on and won't handle changes of the owning Entity name
properly. I'll handle this properly when I get around to the eventual
rewrite of the renderer.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>