Add check(false) to unreachable()
We use unreachable() to indicate a point that should never be reached,
mainly to avoid compiler warnings when it assumes that something can be
reached. However __builtin_unreachable() is defined to have completely
undefined behaviour, so on a debug build it is better to raise an error
if a supposedly unreachable point is reached.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>