Vulkan: Properly implement buffers
We now use device local memory for static buffers, and use staging buffers
to transfer data to them. Additionally, implement map() and unmap(). For
now we don't handle partial maps of buffers without invalidating them, which
would require us to sync with the GPU. This is fine though for our current
uses.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>