Vulkan: Add VulkanRenderPassCompatibilityKey
Vulkan has a number of cases (framebuffers, pipelines) where an object is
created against a specific render pass object, but can be used with any
render pass that is compatible with that one. This type encapsulates the
checking for render pass compatibility, and can be used for caching
objects with this behaviour.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>