Change how a GPURenderTargetDesc must be constructed to refer to the main window
Before the main window was referred to by creating a descriptor with no
attachments. Change it to be a single colour attachment specified by a
null image reference. This makes a bit more sense, and will make things
nicer to handle in the Vulkan backend as it means we can avoid special
casing sections of code for the main window.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>