Change to using Vulkan conventions for clip/NDC coordinates
In Vulkan, the final depth value range is between 0 and 1, whereas in GL
it is between -1 and +1. It is easier to use the Vulkan conventions
primarily and handle the difference in GL. Currently I'm doing this with
ARB_clip_control, but it would also be possible to handle the difference
in shaders using SPIRV-Cross.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>