Use the named parameter idiom to construct simple GPU object descriptors
One of the main points of using descriptors over parameters to object
creation functions is that it is clearer what each parameter is due to
the explicit naming. On the other hand, descriptor structure construction
can look a bit verbose. This commit reduces the verbosity of filling in
descriptors while still giving us the explicit parameter names.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>