Fix objgen construction function generation for VC++
Through the object system we need to be able to construct objects that do
not have public constructors in order to support deserialisation. The
previous solution was to define a lambda which would construct an object
in the scope of the generated staticMetaClass definition, as this was in
the scope of the class and could therefore access its private constructor.

VC++ doesn't like this, and doesn't think that the defined lambda can
access the constructor. Therefore, replace this with the less elegant
solution of defining a static construction method inside the CLASS macro
only when inside the objgen-generated code, and then save a pointer to
that inside the MetaClass.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>