Fix objgen standalone mode
Code generated in standalone mode is intended to be included manually by a
source file that includes the class definitions the code is generated for.
In this case changing the class definition (to add the classConstruct method)
based on whether ORION_OBJGEN_OUTPUT is defined doesn't work, because the
class has already been defined before that was defined. Instead just always
define the classConstruct method, but make it private. This can still be
accessed by the staticMetaClass constructor argument list.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>