Refactor the GPU interface to be based on command lists
Rendering commands within a render pass are now recorded to command lists
which are submitted at the end of the pass. The command list interface allows
command lists to be built in a hierarchy, which will allow for multithreaded
generation of command lists within a render pass.

On Vulkan, the command list interface directly records into secondary command
buffers, which are all submitted onto the per-frame primary command buffer in
a single call to vkCmdExecuteCommands at the end of a pass.

On GL, we use a generic command list implementation which records commands
using an internal representation, which is then replayed and turned into GL
calls at the end of a render pass. In future, replaying the command list
could be offloaded onto a separate thread.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>