Cache vertex data layout objects
Make vertex data layouts cached in the same way that other state objects
are. There will only ever be one object in existence for identical layout
descriptors.

This allows the Vulkan pipeline caching code to use a pointer to a layout
object in its key as it already does for other state objects, which
reduces the cost of pipeline lookups.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>