Move object system from core to engine
I want the core library to be a completely standalone library with no
dependencies on other sections of code, so that it can be used by build
tools, e.g. objgen. Obviously we cannot have objgen depend on a library
which needs to include code that it generates, so move the object system
out to the main engine.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>