Some build system refactoring
Refactor the build to define all our external dependencies, third party
libraries and the engine itself as self-contained "components". Components
can depend on other components, and the build system automatically handles
adding in the necessary include paths and link flags for all required
components. The final application build specifies components it wants to
build against, and all dependencies will be pulled in.

This is in preparation for eventually allowing the engine to be split up
into individual components, rather than having the entire build group all
third party libraries, the whole engine and the application into one
monolithic lump. For now I'll probably only be splitting off the core
library for objgen to make use of, but in future it may be useful to split
out larger sections of the engine. For example, the render_test app is
supposed to be a test for the low level GPU API and does not require the
high level renderer, so if the high level renderer could be entirely cut
out it would allow changes to be made to the low level API and tested
without having to update all of the higher level code initially (this was
something I thought would have been useful while doing the command list
refactoring).

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>