objgen: Don't silently fail when ignore errors option is specified
The reasoning behind the ignore errors flag is that the real compiler will
usually produce much nicer diagnostics than libclang, so we don't cause an
error when libclang raises an error and hope that when we get to properly
compiling the file an error will occur.

However sometimes there can be a problem which happens only when running
objgen, and in that case the error would be silently ignored and then you
would hit mysterious problems at link time. Therefore, even when ignoring
errors, still print them out before carrying on so we can see that objgen
has failed.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>