Clean up filesystem interface
Don't really need the filesystem interface to be implemented as a class
with virtual methods, the implementation (from the platform) is known at
compile time. Also move the code to switch to the engine base directory
to Engine's constructor, with a platform-specific helper. This removes
the need for a filesystem init function entirely. It also makes the
filesystem interface usable for other users of the core library outside
the main engine.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>