Start work on separating game logic and rendering to their own threads
This adds a new "render thread", while the main thread remains as the "game
thread". Nothing is making use of this yet, it's just a draft of how the
2 threads will communicate: the game thread and render thread will have
separate state, the game thread will pass messages (implemented using C++
lambdas) to the render thread to reflect state changes to it. The render
thread runs 1 frame behind the game thread, other than a single point at
the end of a frame there is no synchronisation between them.

Eventually, all rendering (and all direct use of the GPU interface, i.e.
not command list recording) will be required to be done off the render
thread.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>