Move DrawList back out of render_core
This (or something like it) will be used in the high level renderer for
draw call sorting, however it was being (ab)used by other sections of
code to handle the state setup for a draw. Remove these uses of it (and
introduce a helper to Material to make things a bit nicer in the process),
so we can move it back out of render_core.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>