Rearrangments and build changes towards separating the engine into independent components
The build system work is done to allow the "base" engine, without the high
level renderer or physics system, to be built standalone, and a little bit
more work is needed to be able to build the remaining components individually.

It's not possible yet to actually build against only the base engine, because
there are dependencies on the physics system, however these are only for
initialisation.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>