Add a new system for managing per-world state
There are a number of systems which require global per-world state, e.g.
physics and the renderer. This adds a WorldSystem class which acts as a base
for any class managing per-world state. World now has a generic mechanism
to get an instance of WorldSystem-derived classes, creating one the first
time it is requested.

As a result of this, there are no longer any dependencies on the physics
component from the rest of the engine, it is entirely standalone.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>