Don't change resource bindings on an in-use material in the debug overlay
We can't change resource bindings on a material while recording a command list
that references it - on GL, since it replays the command list at submission,
all draws using the material in the command list get only the final state of
the material.

This fixes incorrect textures used in the asset explorer in the debug overlay.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>