Replace asset metadata with serialised loader classes
Previously asset metadata (files sitting alongside an asset containing
parameters to the asset loader class) were just JSON files that used a bunch
of custom parsing code.

Change this to instead make use of the object system. Assets which get
loaded via a loader class (i.e. those which aren't themselves a serialised
object) can now have an extra "<name>.loader" file placed next to them,
which is treated as a serialised AssetLoader-derived class, and is used
instead of creating an AssetLoader from scratch. This can be used to create
loaders with custom parameters as serialised properties.

It is also possible to have a ".loader" file without a real data stream, if
the derived loader class indicates this is allowed. This is used by the
TextureCubeLoader, since it sources its data from a Texture2D for each face
and doesn't need its own data stream.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>