Implement texture views properly
I want to be able to create views with a different format for the next
commit, so may as well do the whole thing properly.

I'm still not overly fond of the implementation (a view is just a
special GPUTexture object). In future I possibly want to move to a D3D-
like model of having entirely separate texture and view objects.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>