Remove special-casing for referring to the main window in the GPU interface
We had special values for GPURenderTargetDesc and GPUTextureImageRef to
refer to the main window, which led to a bunch of places where we had
special cases to do things differently (e.g. get RT properties) based
on whether we were rendering to the main window or a texture.

Improve this by just making the main window have its own GPUTexture
object and use this everywhere, which allows these special cases to be
removed. Any special handling for the main window needed is now handled
internally in the backend.

On Vulkan, we already had a GPUTexture for the main window anyway, so
no change really needed there. On GL, add a dummy object (not backed by
a real texture), which is used to direct render targets/blits to the
backbuffer when seen.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>